+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Add a default contract definition template without a branded logo
 - Add filters for the admin viewset through the field `created_on`
   to search orders on exact date, before date, after date and range of dates.
 - Add filter for admin order viewset by product type